Moved Submit implementation to component and changed to blueprint text field.
